requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.4
decorator==4.4.0
dlib==19.17.0
Flask==1.0.2
future==0.17.1
gast==0.2.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
h5py==2.9.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
networkx==2.2
numpy==1.16.2
parso==0.3.4
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.4.1
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.7.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-resize-image==1.1.18
PyWavelets==1.0.2
PyYAML==5.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
sightengine==1.4.0
six==1.12.0
tensorboard==1.12.2
tensorflow-datasets==1.0.1
tensorflow-gpu==1.12.0
tensorflow-metadata==0.12.1
termcolor==1.1.0
toolz==0.9.0
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
